Rename/move TFA components to swarmpal
