+++ src/swarmpal/toolboxes/tfa/__init__.py
@@ -1,0 +1,4 @@
+from swarmpal.toolboxes.tfa import tfalib
+from swarmpal.toolboxes.tfa.tfa_processor import TfaEfiInputs, TfaMagInputs, TfaProcessor
+
+__all__ = ("TfaMagInputs", "TfaEfiInputs", "TfaProcessor", "tfalib")
