--- src/swarmx/toolboxes/tfa/__init__.py
@@ -1,4 +1,0 @@
-from swarmx.toolboxes.tfa import tfalib
-from swarmx.toolboxes.tfa.tfa_processor import TfaEfiInputs, TfaMagInputs, TfaProcessor
-
-__all__ = ("TfaMagInputs", "TfaEfiInputs", "TfaProcessor", "tfalib")
